requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
alembic==1.12.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6

# Testing dependencies
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
pytest-cov==4.1.0
asgi-lifespan==2.1.0 